sys/src/nix/nix: use $nixdir to specify where the Nix sources are
In order to allow the user to put the nix sources where he sees fit
(for example because one wants to avoid, for now, the sources to be
archived), the initial script to set the name space (sys/src/nix/nix)
uses the variable nixdir that is set, by default, to $home/nix.

README is amended accordingly.

Signed-off-by: Thierry LARONDE <tlaronde@kergis.com>