SEGMAXPG may be different: reintroduced as SEGMAXSIZE (aligned with 9k)
SEGMAXPG was used only once and was defined as SEGMAPSIZE. So I
suppressed it. But I was wrong because even if, for now, it is equal,
it is a distinct notion: in 9k a different value is used, with
SEGMAXSIZE as the name.

So reintroduce it with the 9k name.

Signed-off-by: Thierry LARONDE <tlaronde@kergis.com>