Add the nsec syscall and be compatible with 9legacy
The nsec syscall is added and the Nix added syscalls are defined
after, hence with other numbers, in order for the kernel to be
compatible with existing not Nix userland and in order to segregate
the Nix additions.

The sys/src/libc/9syscall/mkfile is taken as is from 9legacy, dropping
some archs, and adding others, and taking care of special cases about
the syscalls.

Signed-off-by: Thierry LARONDE <tlaronde@kergis.com>